Advanced
Added more features to the calculator, such as the type of calculator (simple or advanced). Where there is more input to enter if the user were to opt for advanced. Added the asset images to the calculator form as well as the arrows on the top and bottom part of the screen and shadows to some of the text